DHCP and DNS config files and install config file created from jinja2 templates.
The DHCP and DNS configuration files are generated using jinja2 templates in Ansible.

This ensures that the MACS, IPs and DNS names are consistent across files.

Many new variables and locals have been defined in libvirt/input-vars.tf to support the rendering of the templates.  Most of these vars are also output vars so they can be used with ansible

The terraform template to create the libvirt VMs also uses the variables to assign MACs.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>