Added ansible tasks to download oc client baremetal installer
Also the pull secret file is copied to the provisioning host, and the install config file is copied into a directory in the provisioning host, which will be used to launch the installation

The command used to download the installer is saved in a file in the provisioning host for reference

The number of workers variable in terraform libvirt is validated to make sure that the value used is withing parameters

A new input variable for terraform libvirt is created to define the OCP version to deploy

Quotes are stripped from the cluster name is case they were used in the command line definition of the variable.  And that stripped value is the one used in the output variable.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>