User and password to access sushy tools and vbmc are centrally generated
The user is defined in variable manged_user, default is admin but now can be changed.
The password is automatically generated by ansible in a task that is common no matter if redfish or vbmc is selected.
Both user and password definitions are the same no matter whether sushy or vbmc is used.

Since the password is generated in one play and then used in different plays, the value must be accessed using hostvars[]

The ansible tasks to remove ssh know hosts stale entries is properly defined now

Updating the variables for support and provision groups in inventory file is using the correct reference

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>