Support for EC2 spot instances has been added.
Depending on the value of the terraform input var spot_instance, the EC2 instance created is a standard or spot instance.  The spot instance does not support RHEL 8.5 at the moment, so 8.4 is used instead.

The count trick is used to conditionally create the spot or the standard resources in terraform.

Python pip3 is installed since it is not present in a RHEL 8.4.  This wasn't necessary when using RHEL 8.5

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>