Automatic creation of bonding NIC if requested by user.
The user can request the creation of a bonding NIC for master and nodes via the variable bonding_nic.  OCP version must be 4.10+

Maximum number of worker nodes increased to 128

The vbmcd and sushy services are restarted in ansible instead of started to make sure any change in configuration is applayed on every playbook execution.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>