The terraform variable resources_id is added to add a user defined string to the names of AWS resources.
With the adition to a user defined string to AWS resources it is easier to identify the resources belonging to a particular project.  If the user does not provide a value for the variable resources_id a random string will be generated.

The terraform variable vpc_name that allowed giving a name to the VPC created in AWS has been removed, it was not very useful.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>