Updated to OCP 4.11 and aws terraform provider 4.28.0
The EC2 metal instance uses RHEL 9 as OS, the reason being that sushy
tools cannot be installed with the pip version included in 8.6 anymore,
and updating the pip version causes trouble with other python modules
installed as rpm packages.

The provision and support KVM VMs still use RHEL 8.6 because the
openshift installer does not support RHEL 9 for the provision host, the
support host could use RHEL 9 but for configuration simplicity both
use the sema OS version.

The baremetal host tasks in the support_setup.yaml ansible playbook are run
before the support host tasks, so if anything fails with the support
host tasks, the VMs can be checked using virt-manager.

The cpu mode for the support VM uses the host-passthrough mode as the
rest of VMs, otherwise the VM does not boot in RHEL 9

Signed-off-by: Jose Ignacio Jerez Rodriguez <jose.jerez@redhat.com>