Provision VM gets a static IP
The provision VM has been configured to get a static IP instead of a
dynamic one from DHCP, this is because if the dhcp service in the
support VM was started a long time after the provision VM was running,
the provision VM may have given up requesting an IP from DHCP.

The static IP, gateway and nameserver for the provision VM is defined in the cloud init data

Removed config section to provide an IP for the provision VM from the DHCP config file

The hostname for the provisiob VM is set up by ansible.  This
information was provided by the DHCP server before.

tmux is installed in the AWS metal instance

Updated documentation

Signed-off-by: Jose Ignacio Jerez Rodriguez <jose.jerez@redhat.com>